+++ src/linear_curve.rs
@@ -12,7 +12,7 @@ pub struct LinearCurve {
     pub vertex_buffer: Buffer<[f32; 4]>,
     pub index_buffer: Buffer<u32>,
     pub flag_buffer: Buffer<u32>,
-    pub normal_buffer: Option<Buffer<[u32; 3]>>,
+    pub normal_buffer: Option<Buffer<[f32; 3]>>,
 }
 
 impl LinearCurve {
